+++ vpop_calibration/test/saem/test_mstep.py
@@ -5,12 +5,12 @@ import torch
 
 def test_mstep():
     nb_pdu = 3
-    indiv_design_matrix = torch.eye(nb_pdu).float()
+    indiv_design_matrix = torch.eye(nb_pdu).double()
     nb_chains = 1
     nb_patients = 2
     design_matrix = torch.stack((indiv_design_matrix, indiv_design_matrix), dim=0)
 
-    gaussian_params = torch.zeros((nb_chains, nb_patients, nb_pdu)).float()
+    gaussian_params = torch.zeros((nb_chains, nb_patients, nb_pdu)).double()
 
     state = MStepState.from_init_gaussian_params(
         design_matrix=design_matrix,
@@ -20,18 +20,18 @@ def test_mstep():
         init_gaussian_params=gaussian_params,
     )
 
-    new_gaussian_params = torch.ones((nb_chains, nb_patients, nb_pdu)).float()
+    new_gaussian_params = torch.ones((nb_chains, nb_patients, nb_pdu)).double()
     proposal = state.update(new_gaussian_params=new_gaussian_params, learning_rate=0.1)
 
 
 def test_state_dict():
     nb_pdu = 3
-    indiv_design_matrix = torch.eye(nb_pdu).float()
+    indiv_design_matrix = torch.eye(nb_pdu).double()
     nb_chains = 1
     nb_patients = 2
     design_matrix = torch.stack((indiv_design_matrix, indiv_design_matrix), dim=0)
 
-    gaussian_params = torch.zeros((nb_chains, nb_patients, nb_pdu)).float()
+    gaussian_params = torch.zeros((nb_chains, nb_patients, nb_pdu)).double()
 
     state = MStepState.from_init_gaussian_params(
         design_matrix=design_matrix,
